Address review feedback: move exception, javadoc private methods, self-clearing resolveAll
- Move UnlockConditionParseException to us.eunoians.mcrpg.exception package
- Add Javadoc to private methods resolveUncached() and getUnlockConditionsSection()
- Make resolveAll() clear the cache first so it's self-contained
- Document the underscore-to-hyphen replacement in getUnlockConditionsSection()
