+++ src/test/java/us/eunoians/mcrpg/ability/unlock/builtin/AllOfUnlockConditionTypeTest.java
@@ -6,7 +6,7 @@ import org.bukkit.NamespacedKey;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import us.eunoians.mcrpg.McRPGBaseTest;
-import us.eunoians.mcrpg.ability.unlock.UnlockConditionParseException;
+import us.eunoians.mcrpg.exception.UnlockConditionParseException;
 import us.eunoians.mcrpg.ability.unlock.UnlockConditionType;
 import us.eunoians.mcrpg.entity.holder.AbilityHolder;
 import us.eunoians.mcrpg.entity.holder.SkillHolder;
